requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
colorama==0.4.3
more-itertools==8.0.2
packaging==19.2
pluggy==0.13.1
py==1.8.0
pyparsing==2.4.5
pytest==5.3.2
six==1.13.0
wcwidth==0.1.7
